listening circle completed
I got impatient and rushed ahead. it was fun to see the final output but there was a lot from the instructor i didn't understand.

i'm happy to have done this so I can hopefully reverse engineer it with the other things I'm learning, but I am keen to go back to where I'm supposed to be.